refactor: split ContactList into smaller components
