+++ apps/next/src/pages/Contacts/components/index.ts
@@ -2,3 +2,5 @@ export * from './AddressField';
 export * from './InvisibleInput';
 export * from './ContactNameField';
 export * from './RemoveContactWarning';
+export * from './EmptyContactList';
+export * from './SearchableContactList';
